requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
atomicwrites==1.4.1
attrs==25.4.0
certifi==2024.07.04
chardet==5.2.0
colorama==0.4.6
exceptiongroup==1.0.4
Faker==37.12.0
greenlet==3.2.4
idna==3.7
iniconfig==2.1.0
jsonpath==0.82.2
loguru==0.7.3
more-itertools==10.8.0
packaging==25.0
playwright==1.55.0
pluggy==1.5.0
py==1.11.0
pyee==13.0.0
PyMySQL==1.1.2
pytest==7.4.4
pytest-base-url==2.0.0
pytest-playwright==0.7.1
python-dateutil==2.8.2
python-slugify==7.0.0
PyYAML==6.0.3
requests==2.32.5
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
typing-extensions==4.13.2
urllib3==1.26.13
win32-setctime==1.1.0
xlrd==1.2.0